Bug #34221: 96_samba4_sid_allocation: remove redundant rid check
This commit removes a redundant `udm_rid != next_rid` in the test case. This is
not necessary as this check is implicitly included in the previous `ldap_sid !=
samba_sid` check.

Also removes the function `get_samba_rid_via_udm()` as it was only used in the
removed code chunk.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-school-4.1r2@75518 1b283449-9f2f-0410-b571-8eb228e1a901