Bug #44255: fixed regression when creating exam objects
The exam-master module tried to create a UDM object via the following
(suboptimal) code:

ucs-school-4.2/ucs-school-umc-exam/umc/python/schoolexam-master/__init__.py:
94  self._examGroup = module_groups_group.object(None, ldap_admin_write, position, examGroup)
95  self._examGroup.open()
96  self._examGroup['name'] = examGroupName
97  self._examGroup['sambaGroupType'] = self._examGroup.descriptions['sambaGroupType'].base_default[0]
98  self._examGroup.create()

"examGroup" contained the DN of the group object to be created. But the object does not exist yet.
Before UCS 4.2, object() didn't raise an exception if the given DN does not exist. Now a noObject exception is raised.
The correct line would not specify the DN:

94  self._examGroup = module_groups_group.object(None, ldap_admin_write, position)

The same applies to line 118:

118  exam_user_container = module_containers_cn.object(None, ldap_admin_write, position, examUsers)

vs.

118  exam_user_container = module_containers_cn.object(None, ldap_admin_write, position)

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.2/ucs-school-4.2@78693 1b283449-9f2f-0410-b571-8eb228e1a901