Bug #45352: update during prerun() and on special group changes
The list of local schools is now updated via prerun() and if a UCS@school
group with groupname suffix "-Edukativnetz" is modified/created. This way
the gap between prerun() and possible new OUs or permission changes is
reduced to zero with a very small overhead (LDAP lookups if special groups
are altered).