Bug #45473: port feature ucsschool/exam/group/ldap/blacklist/regex from 4.1R2 to 4.2
In some customer environments, school users are members of several groups
that are useless for exam users. To speed up the start of the exam mode
and to reduce the number of LDAP group changes, a regular expression for
group dns may be specified, which are not taken over to exam users.