Bug #43980: increase exam start timeout from 5min to 30min
During exam start, the UMC module waits for all exam users to be replicated
to the school slave. Depending on the number of exam users and the LAN/WAN
connection to the UCS master, this may take several minutes.
In some customer environments, the old limit of 5 minutes has been reached
in normal conditions. Therefore the timeout has been raised to 30mins.