Bug #46453: add dependency to ucs-school-import
IIRC we started in Bug 30499 using ucs_registerLDAPSchema(). Unfortunately I
have no idea why we decided to use ucs_registerLDAPSchema() in the join script
of ucs-school-import instead of ucs-school-import-schema. Maybe the idea was
to remove ucs-school-import-schema entirely on the long run.

Currently the schema is installed in the *filesystem* via
ucs-school-import-schema. ucs-school-import depends on
ucs-school-import-schema, so the schema file is in the filesystem when the
ucs-school-import.postinst is called. Within this postinst the joinscript
35ucs-school-import.inst is called directly and the schema is registered
by the joinscript.

ucs-school-ldap-acls-master stores the LDAP ACLs in the *filesystem* upon
package extraction. The postinst of ucs-school-ldap-acls-master also calls
its joinscript 70ucs-school-ldap-acls-master.inst directly where the ACLs
are registered in LDAP.

If the joinscripts are called by univention-join, the correct order is
preserved. First 35ucs-school-import.inst then
70ucs-school-ldap-acls-master.inst.

But since both postinst scripts call their joinscript directly on a UCS
master, the registration order of ACLs and schema depends on the execution
order of those two postinst scripts.

A dependency in ucs-school-ldap-acls-master will be added that requires
ucs-school-import to be installed and configured first.