Bug #48068: fixed logic bug in create_school_admin()
Without this fix, the created school admin is only member in the OU
specified by "ou_name". The values in "schools" are completely ignored.
With this patch, "ou_name" is used if "schools" is not defined.
In any case, the user is created below the OU specified by "ou_name".