Bug #50274: Adjust 55_class_assignment.py to changed ACL
The test 55_class_assignment.py was written while the school ldap master
acl were active on a school slave.
That prevented a school admin from ou A to read anything from ou B even
if both ous were on the same slave DC. With the removal of the master
school ldap acl from the slave, all ous on the same slave DC can be read
by all users on that DC.

The test has been adjusted to created ous on different DCs.