Bug #50772: The password dialog for users using saml has been removed
The password was only required for a local umc connection which was used
to control the computerroom. I slightly refactored the computerroom
module to use it directly from from the exam module. This removed the
need to know the users plain text password.
This has been done for more stability when using saml and should provide
an overall better user experience.