Bug #54120: fix 96_samba4_sid_allocation.py
The rIDNextRID attribute seems only to be set after the first RID has
been requested. In some test scenarios no object that needed a rid is
ever created through samba and the attribute is missing.
A rIDNextRID attribute that stays missing should be the same as a non
changing rIDNextRID attribute.