Bug #54040: hide school selection in web UI
Don't show the school input in the popup to add new users to a
workgroup. This way the user won't be tempted to add users from
other schools (which would fail anyway). The input is kept because its
value is populated by an async function and then used by other widgets.