Bug #55104: Allow manual docs rebuild
I think it is useful to be able to trigger this pipeline manual. For
example if the build dependecies changed.

But to be honest I added this so the build stage doesn't fail if the
documentation didn't change.
gitlab throws an error if no pipeline did run in a stage.