Bug #55104: Mark manual docs pipeline as green
`alow_failure: true` doesn't mark the pipeline as blocked, even if the
job is never triggered. Since this is used to trigger a rebuild of
the documentation that should be ok. This shoudl result in cleaner
pipeline results.

The pipeline result is now green even if the build stage did not run any
job.