Issue #958: run pre-commit hooks in the same python environment
This commit adds a default python version of 3.7 to all tests. This is
especially important for black, flake8, and bandit, which should check
under the python version used by our servers.

For pymarkdown, only versions >= 3.8 are supported. Since this is just
checking markdown and not python files, this should be safe to run under
a different version than the rest of the tests. We're explicitly
specifying python3.8, because it's the python version currently used by
our test server docker image.

NOTE: although pre-commit has a `default_language_version` option that
can be specified for all hooks, this feature does not work as expected,
hence why it wasn't used. If a repo such as pymarkdown has their own
pre-commit version specified, pre-commit ignores the univention default
setting in favor of the repo's setting. The only way to prevent hooks
from running under an unexpected python version is to specify the
`language_version` manually for each hook.