Issue #961: Bug #55768: allow colons in exam names
Teachers who try to run exams with colons in the names, will receive
validation error messages because the role string for the exam follows
the format "student:exam:{exam_name}". If there are colons in the exam
name, splitting the role string will yield more than the expected number
of results and the role becomes invalid.

The solution is to replace colons in the exam name with a string that is
unlikely to be used by other exams, when generating the role string. For
this, we chose "&&&".