Issue #1031: ignore tracebacks for importer test
In ucs-test there is a test, `99_end/01_var_log_tracebacks.py` that
checks to see if there are any tracebacks in the log file, and if so,
fails. We found that `90_ucsschool/302_http-api_class_column_no_school`
logs ValidationError tracebacks, but these tracebacks are an integral
part of the test, so should be ignored by the ucs-test.

In order to ignore the ValidationErrors, we need a regex that is easy to
distinguish from ValidationErrors that might be thrown by other tests.
For this, we're creating a custom OU that is used by the test,
`302_http_api_no_school`, so it is easy to ignore these tracebacks.