Bug #56152: Add new UCR variable for username validation
With the resolution of Bug #53519 a new validation was added for usernames
which tests for compliance with the Windows naming conventions. As there
was no way of disabling this validation, it was no longer possible to do an
import if non-compliant user names are present.

This commit introduces:

- A UCR variable ("ucsschool/validation/username/windows-check") with which
  administrators can manually enable or disable the new validation
- A check in the joinscript which automatically enables the this UCR
  variable if all present usernames are compliant and the variable is
  not already set
- A script which is used in the joinscript which can be manually used by
  Administrators to list all non-compliant usernames

References:
- https://git.knut.univention.de/univention/ucsschool/-/issues/1050
- https://forge.univention.org/bugzilla/show_bug.cgi?id=56152

Bug #56152