Bug #53584: Rework section about `import_networks`
The section was ambiguous about what benefits brings the script in
contrast to creating computer accounts manually. The secret sauce is the
DHCP policy. Computer accounts create network objects that usually don't
have a DHCP policy assigned to them.

The script `import_networks` creates the network objects and
additionally a DHCP policy and assigns it to the network object. The
DHCP policy configures the gateway, the DNS server and the WINS server.