fix(ucs-school-import): Fix performance regression
Bug #57408

The `csv.Sniffer` class is used with a sample of unbounded size.
As the resulting calculations are quadratic in complexity
(`regex.findall`), this leads to long startup times.
On a default VM, for 10000 users, the time to detect the csv dialect
takes 60 seconds, for 20000 it is already 240 seconds and for 100000
users it will be approximately 100 minutes.