fix: Change ownership only if necessary
It is possible that the immutable flag is set for the data
directory in the context of an exam. We now check if the correct
owner and group is set and try to change it if not. If the immutable
flag is set and the incorrect owner and group are set, we log an error.

Issue univention/ucsschool#1301
Bug #57661