ci: allow downgrades
If a push with package changes happens on the main branch slightly
before/while this pipeline is being started, the pipeline will likely
fail due to the attempt of a package downgrade.

Since we have to run the pipeline in order to perform a merge of the
feature branch, it is no problem to allow downgrades within the
pipeline. After the necessary rebase the pipeline has to be restarted
and therefore will run with the latest version.

Issue univention/ucsschool#1362