feat(u-s-u-computerroom): set the width of screenshot view to 100%
To be able to show more screenshots on larger screens, the width of
the screenshot view with its grid will automatically change to 100%.
The ScreenshotView is designed to occupy the full width of its
container. To achieve this, the parent element with the CSS class
"umcModuleWrapper" needs to have its width set to 100%. However, this
should only occur when the "screenShotView__screenshotContainer" element
is visible, as "umcModuleWrapper" is used by multiple UMC modules.

This commit adds a JavaScript function, `styleUmcModuleWrapper`, which
checks the visibility of the
"screenShotView__screenshotContainer" element and sets the width of its
"umcModuleWrapper" parent accordingly.

Additionally, a MutationObserver is implemented to update the CSS class
whenever the visibility of elements within the ScreenshotView changes.
This ensures that the layout remains correct even after switching
between UMC modules or within the UMC module.

Issue univention/ucsschool#1404