test: fix missing UMC call to end an exam
If computerroom/exam/finish is not called after schoolexam/exam/finish,
the computerroom's info file is not updated and the room remains blocked
for further exams. The ucs-test lib was missing this call.
And to sucessfully call computerroom/exam/finish the room has to
be aquired first.

Issue univention/ucsschool#1408