feat: Add import job cleanup step
Introduced the UCRV `ucsschool/import/http_api/import_jobs_to_keep` with
which the number of import jobs to keep in the database can be configured.
At the end of each import job, a cleanup is performed: If there are more
jobs than configured, the oldest jobs are deleted. During the deletion
process, files related to the deleted import job are removed as well.

Bug #52173

Issue univention/ucsschool#293